More code refactors
