+++ src/components/elements/calendar/CalendarDate.tsx
@@ -1,38 +1,17 @@
 "use client";
 
-import { Tab } from "@headlessui/react";
 import { clsx } from "clsx";
-import { format, isSameMonth, isToday, parseISO } from "date-fns";
-import { Form, Formik } from "formik";
-import { AnimatePresence, MotionConfig, Variants, motion } from "framer-motion";
+import { format, isSameMonth, isToday } from "date-fns";
+import { MotionConfig } from "framer-motion";
 import { useState } from "react";
 import { HiX } from "react-icons/hi";
 import useMeasure from "react-use-measure";
 
-import {
-  AllDayEventFormValues,
-  MultiDayEventFormValues,
-  NormalEventFormValues,
-  YupSchemas,
-} from "@/app-constants";
-import {
-  AnimationWrapper,
-  Event,
-  Modal,
-  animations,
-  DatePicker,
-  DatePickerWithRange,
-  TimePicker,
-  FormikInput,
-  ResizablePanel,
-  AllDayEventCalendarDisplay,
-  ThreeElementMovingBox,
-} from "@/components";
+import { AnimationWrapper, Event, Modal, ResizablePanel, animations } from "@/components";
 import { useUser } from "@/hooks";
-import { cn } from "@/lib";
 import { EventData } from "@/types";
 
-import { CalendarEventTab } from ".";
+import { EventCreationTabs } from ".";
 
 interface Props {
   events: EventData[];
@@ -42,48 +21,12 @@ interface Props {
   isAnimating: boolean;
 }
 
-const duration = 0.4;
-
-const eventContentVariants = (toLeft: boolean) => {
-  const variant: Variants = {
-    hidden: {
-      x: toLeft ? 20 : -20,
-      opacity: 0,
-    },
-    active: {
-      x: 0,
-      opacity: 1,
-      transition: {
-        duration,
-      },
-    },
-  };
-  return variant;
-};
-
 const transition = { type: "ease", ease: "easeInOut", duration: 1 };
 
 export const CalendarDate = ({ events, date, month }: Props) => {
   const [ref, bounds] = useMeasure();
   const { data: user } = useUser();
-  const [selectedIndex, setSelectedIndex] = useState(0);
   const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const [normalEventInitialValues] = useState<NormalEventFormValues>({
-    start: "",
-  });
-  const [allDayEventInitialValues] = useState<AllDayEventFormValues>({
-    title: "",
-    subTitle: "Treeninguid ei toimu",
-    start: "",
-  });
-  const [multiDayEventInitialValues] = useState<MultiDayEventFormValues>({
-    title: "",
-    start: "",
-    end: "",
-  });
-
-  const [time, setTime] = useState<Date>(new Date());
 
   const openModal = () => {
     setIsModalOpen(true);
@@ -150,172 +93,7 @@ export const CalendarDate = ({ events, date, month }: Props) => {
                 </div>
               </div>
               <div className="p-10">
-                <Tab.Group selectedIndex={selectedIndex} onChange={setSelectedIndex}>
-                  <Tab.List
-                    className={clsx(
-                      "flex flex-row relative px-1 w-full flex-1 mb-6 bg-stone-100 rounded-xl max-w-md m-auto md:mb-12"
-                    )}
-                  >
-                    <CalendarEventTab selectedIndex={selectedIndex} index={0}>
-                      Trenn
-                    </CalendarEventTab>
-                    <CalendarEventTab selectedIndex={selectedIndex} index={1}>
-                      Päevane sündmus
-                    </CalendarEventTab>
-                    <CalendarEventTab selectedIndex={selectedIndex} index={2}>
-                      Mitme päevane sündmus
-                    </CalendarEventTab>
-                    <ThreeElementMovingBox selectedIndex={selectedIndex} />
-                  </Tab.List>
-                  <Tab.Panels>
-                    <div className="p-3">
-                      <AnimatePresence>
-                        {/* normal */}
-                        <Tab.Panel
-                          as={motion.div}
-                          initial="hidden"
-                          animate="active"
-                          exit="exit"
-                          variants={eventContentVariants(true)}
-                          key={"panel 1"}
-                        >
-                          <Formik
-                            initialValues={normalEventInitialValues}
-                            validationSchema={YupSchemas.Events.NormalEvent}
-                            validateOnChange={true}
-                            onSubmit={(values, { setSubmitting }) => {
-                              setSubmitting(true);
-
-                              setSubmitting(false);
-                            }}
-                          >
-                            {() => {
-                              return (
-                                <Form>
-                                  <TimePicker time={time} onChange={setTime} />
-                                  <DatePicker />
-                                </Form>
-                              );
-                            }}
-                          </Formik>
-
-                          {/* user chooses groups, time, trailers, title */}
-                          {/* add presets */}
-                          {/* add pick for multiple days */}
-                        </Tab.Panel>
-
-                        {/* all day */}
-                        <Tab.Panel
-                          as={motion.div}
-                          initial="hidden"
-                          animate="active"
-                          exit="exit"
-                          variants={eventContentVariants(false)}
-                          key={"panel 2"}
-                        >
-                          <Formik
-                            initialValues={allDayEventInitialValues}
-                            validationSchema={YupSchemas.Events.AllDayEvent}
-                            validateOnChange={true}
-                            onSubmit={(values, { setSubmitting }) => {
-                              setSubmitting(true);
-
-                              setSubmitting(false);
-                            }}
-                          >
-                            {({ values }) => {
-                              return (
-                                <Form>
-                                  <DatePicker />
-                                  <FormikInput
-                                    required
-                                    className="w-full"
-                                    label="Pealkiri"
-                                    placeholder="Pealkiri"
-                                    name="title"
-                                  />
-                                  <FormikInput
-                                    required
-                                    className="w-full"
-                                    label="Ala pealkiri"
-                                    placeholder="Ala pealkiri"
-                                    name="subTitle"
-                                  />
-
-                                  <div
-                                    className={cn(
-                                      "h-52 w-36 border-1 mt-2 border-stone-100 border-t-0 flex flex-col"
-                                    )}
-                                  >
-                                    <p className="self-center font-semibold justify-self-center text-stone-500">
-                                      {values.start !== ""
-                                        ? format(parseISO(values.start), "EEEE")
-                                        : "Esmaspäev"}
-                                    </p>
-                                    <div className="flex flex-col items-center justify-center">
-                                      <div className="text-xs font-medium font-number sm:text-sm md:text-base text-text-primary">
-                                        {values.start !== ""
-                                          ? parseISO(values.start).getDay()
-                                          : "1"}
-                                      </div>
-                                    </div>
-                                    <div className="flex flex-col justify-center flex-grow text-center">
-                                      <AllDayEventCalendarDisplay
-                                        title={values.title}
-                                        sub_title={values.subTitle}
-                                      />
-                                    </div>
-                                  </div>
-                                </Form>
-                              );
-                            }}
-                          </Formik>
-
-                          {/* title and subtitle */}
-                          {/* add presets */}
-                        </Tab.Panel>
-                        {/* multi day */}
-                        <Tab.Panel
-                          as={motion.div}
-                          initial="hidden"
-                          animate="active"
-                          exit="exit"
-                          variants={eventContentVariants(false)}
-                          key={"panel 3"}
-                        >
-                          <Formik
-                            initialValues={multiDayEventInitialValues}
-                            validationSchema={YupSchemas.Events.MultiDayEvent}
-                            validateOnChange={true}
-                            onSubmit={(values, { setSubmitting }) => {
-                              setSubmitting(true);
-
-                              setSubmitting(false);
-                            }}
-                          >
-                            {() => {
-                              return (
-                                <Form>
-                                  <FormikInput
-                                    required
-                                    className="w-full"
-                                    label="Pealkiri"
-                                    placeholder="Pealkiri"
-                                    name="title"
-                                  />
-                                  <DatePickerWithRange />
-                                </Form>
-                              );
-                            }}
-                          </Formik>
-
-                          {/* title */}
-                          {/* add presets */}
-                        </Tab.Panel>
-                      </AnimatePresence>
-                    </div>
-                  </Tab.Panels>
-                </Tab.Group>
+                <EventCreationTabs />
               </div>
             </ResizablePanel>
           </div>
