+++ src/components/elements/calendar/event-creation/index.ts
@@ -1,2 +1,3 @@
+export * from "./EventCreationTabs";
 export * from "./CalendarEventTab";
 export * from "./time-date-pickers";
