+++ src/components/elements/calendar/event-creation/time-date-pickers/DatePicker.tsx
@@ -1,9 +1,10 @@
 "use client";
 
 import { format } from "date-fns";
-import { Calendar as CalendarIcon } from "lucide-react";
 import { useState } from "react";
+import { HiOutlineCalendar } from "react-icons/hi";
 
+import { AnimationWrapper, animations } from "@/components";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/elements/Popover";
 
 import { DatePickerCalendar, DatePickerCalendarButton } from ".";
@@ -13,11 +14,16 @@ export function DatePicker() {
 
   return (
     <Popover>
-      <PopoverTrigger asChild>
-        <DatePickerCalendarButton variant="outline" className="justify-start font-normal text-left">
-          <CalendarIcon className="w-4 h-4 mr-2" />
-          {date ? format(date, "PPP") : <span>Vali kuupäev</span>}
-        </DatePickerCalendarButton>
+      <PopoverTrigger>
+        <AnimationWrapper variants={animations.subtleScale}>
+          <DatePickerCalendarButton
+            variant="ghost"
+            className="items-center justify-center text-left group"
+          >
+            <HiOutlineCalendar className="w-6 h-6 mr-2 text-stone-700 group-hover:text-stone-800" />
+            {date ? format(date, "PPP") : <p>Vali kuupäev</p>}
+          </DatePickerCalendarButton>
+        </AnimationWrapper>
       </PopoverTrigger>
       <PopoverContent className="w-auto p-0 z-[5001]">
         <DatePickerCalendar mode="single" selected={date} onSelect={setDate} initialFocus />
