+++ src/components/elements/calendar/event-creation/time-date-pickers/DatePickerCalendar.tsx
@@ -62,8 +62,8 @@ const DatePickerCalendar = ({
         nav_button_previous: "absolute left-1",
         nav_button_next: "absolute right-1",
         table: "w-full border-collapse space-y-1",
-        head_row: "flex",
-        head_cell: "text-stone-700 rounded-md w-9 font-normal text-[0.8rem]",
+        head_row: "flex text-stone-700",
+        head_cell: "text-stone-700 rounded-md w-9 text-[0.8rem]",
         row: "flex w-full mt-2",
         cell: "text-center text-sm p-0 relative first:[&:has([aria-selected])]:rounded-l-md last:[&:has([aria-selected])]:rounded-r-md focus-within:relative focus-within:z-20",
         day: cn(
