+++ src/components/elements/calendar/event-creation/time-date-pickers/DateRangePicker.tsx
@@ -1,10 +1,11 @@
 "use client";
 
 import { format } from "date-fns";
-import { Calendar as CalendarIcon } from "lucide-react";
 import { useState } from "react";
 import { DateRange } from "react-day-picker";
+import { HiOutlineCalendar } from "react-icons/hi";
 
+import { AnimationWrapper, animations } from "@/components";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/elements/Popover";
 
 import { DatePickerCalendar, DatePickerCalendarButton } from ".";
@@ -14,25 +15,27 @@ export const DatePickerWithRange = () => {
 
   return (
     <Popover>
-      <PopoverTrigger asChild>
-        <DatePickerCalendarButton
-          id="date"
-          variant="outline"
-          className="justify-start font-normal text-left"
-        >
-          <CalendarIcon className="w-4 h-4 mr-2" />
-          {date?.from ? (
-            date.to ? (
-              <>
-                {format(date.from, "LLL dd, y")} - {format(date.to, "LLL dd, y")}
-              </>
+      <PopoverTrigger>
+        <AnimationWrapper variants={animations.subtleScale}>
+          <DatePickerCalendarButton
+            id="date"
+            variant="ghost"
+            className="items-center justify-center text-left group"
+          >
+            <HiOutlineCalendar className="w-6 h-6 mr-2 text-stone-700 group-hover:text-stone-800" />
+            {date?.from ? (
+              date.to ? (
+                <>
+                  {format(date.from, "LLL dd, y")} - {format(date.to, "LLL dd, y")}
+                </>
+              ) : (
+                format(date.from, "LLL dd, y")
+              )
             ) : (
-              format(date.from, "LLL dd, y")
-            )
-          ) : (
-            <span>Vali kuupäev</span>
-          )}
-        </DatePickerCalendarButton>
+              <p>Vali kuupäev</p>
+            )}
+          </DatePickerCalendarButton>
+        </AnimationWrapper>
       </PopoverTrigger>
       <PopoverContent className="w-auto p-0 z-[5001]" align="start">
         <DatePickerCalendar
